+++ src/controls.rs
@@ -7,6 +7,7 @@ pub(crate) mod menu_handler;
 pub(crate) mod paint_router;
 pub(crate) mod panel_handler;
 pub(crate) mod progress_handler;
+pub(crate) mod richedit_handler;
 pub(crate) mod splitter_handler;
 pub(crate) mod styling_handler;
 pub(crate) mod treeview_handler;